+++ projects/mk-kit/src/lib/components/charts/index.ts
@@ -4,3 +4,5 @@ export * from './bar-chart';
 export * from './line-chart';
 export * from './donut-chart';
 export * from './gauge';
+export * from './scatter-chart';
+export * from './heatmap';